Header/Start page set up per Figma. Blobs might need repositioning for aesthetic reasons, but all is in place.
