+++ src/components/layout/Footer.tsx
@@ -19,7 +19,7 @@ import Medium from '../../images/icons-footer/social_medium.png';
 
 const Footer = () => (
   <FooterNavbar
-    className="justify-content-center nav-container"
+    className="container fixed-bottom justify-content-center nav"
     expand="sm"
     sticky="bottom"
   >
