+++ src/styles/App.scss
@@ -1,1 +1,6 @@
 @import "custom";
+
+html, body, main, #root, #app, #app > div {
+  width: 100%;
+  overflow-x: hidden;
+}
